+++ python/cudf/cudf/core/reshape.py
@@ -3,7 +3,6 @@ import itertools
 
 import numpy as np
 import pandas as pd
-
 import cudf
 
 _axis_map = {0: 0, 1: 1, "index": 0, "columns": 1}
@@ -15,7 +14,8 @@ def _align_objs(objs, how="outer"):
     Parameters
     ----------
     objs : list of DataFrame, Series, or Index
-
+    how : How to handle indexes on other axis (or axes),
+    similar to join in concat
     Returns
     -------
     A bool for if indexes have matched and a set of
@@ -62,7 +62,7 @@ def _normalize_series_and_dataframe(objs, axis):
             objs[idx] = o.to_frame(name=name)
 
 
-def concat(objs, axis=0, ignore_index=False, sort=None):
+def concat(objs, axis=0, join="outer", ignore_index=False, sort=None):
     """Concatenate DataFrames, Series, or Indices row-wise.
 
     Parameters
@@ -70,6 +70,8 @@ def concat(objs, axis=0, ignore_index=False, sort=None):
     objs : list of DataFrame, Series, or Index
     axis : {0/'index', 1/'columns'}, default 0
         The axis to concatenate along.
+    join : {'inner', 'outer'}, default 'outer'
+        How to handle indexes on other axis (or axes).
     ignore_index : bool, default False
         Set True to ignore the index of the *objs* and provide a
         default range index instead.
@@ -150,6 +152,17 @@ def concat(objs, axis=0, ignore_index=False, sort=None):
     0      c       3    cat
     1      d       4    dog
 
+    Combine ``DataFrame`` objects with overlapping columns
+    and return only those that are shared by passing ``inner`` to
+    the ``join`` keyword argument.
+
+    >>> cudf.concat([df1, df3], join="inner")
+      letter  number
+    0      a       1
+    1      b       2
+    0      c       3
+    1      d       4
+
     Combine ``DataFrame`` objects horizontally along the
     x axis by passing in ``axis=1``.
 
@@ -169,7 +182,6 @@ def concat(objs, axis=0, ignore_index=False, sort=None):
         raise ValueError("No objects to concatenate")
 
     objs = [obj for obj in objs if obj is not None]
-
     # Return for single object
     if len(objs) == 1:
         if ignore_index:
@@ -177,6 +189,8 @@ def concat(objs, axis=0, ignore_index=False, sort=None):
                 data=objs[0]._data.copy(deep=True),
                 index=cudf.RangeIndex(len(objs[0])),
             )
+            if axis == 1:
+                result.columns = pd.RangeIndex(len(result.columns))
         else:
             result = objs[0].copy()
         return result
@@ -213,14 +227,28 @@ def concat(objs, axis=0, ignore_index=False, sort=None):
     # when axis is 1 (column) we can concat with Series and Dataframes
     if axis == 1:
 
-        assert typs.issubset(allowed_typs)
+        if not typs.issubset(allowed_typs):
+            raise TypeError(
+                "Can only concatenate Series and DataFrame objects when axis=1"
+            )
         df = cudf.DataFrame()
         _normalize_series_and_dataframe(objs, axis=axis)
 
-        objs, match_index = _align_objs(objs)
+        old_objs = objs
+        objs = [obj for obj in objs if obj.shape != (0, 0)]
+        if len(objs) == 0:
+            return cudf.DataFrame()
+        empty_inner = False
+        if join == "inner":
+            # don't filter out empty df's
+            if any(obj.empty for obj in old_objs):
+                empty_inner = True
 
+        objs, match_index = _align_objs(objs, how=join)
         for idx, o in enumerate(objs):
-            if not ignore_index and idx == 0:
+            # if join is inner the index remains unchanged
+            # and (not ignore_index or join=='inner')
+            if idx == 0 and not ignore_index:
                 df.index = o.index
             for col in o._data.names:
                 if col in df._data:
@@ -235,14 +263,24 @@ def concat(objs, axis=0, ignore_index=False, sort=None):
         for o in objs[1:]:
             result_columns = result_columns.append(o.columns)
 
-        df.columns = result_columns.unique()
         if ignore_index:
-            df.index = cudf.RangeIndex(len(objs[0]))
-            return df
-        elif not match_index:
+            # with ignore_index the column names change to numbers
+            df.columns = pd.RangeIndex(len(result_columns.unique()))
+        if empty_inner:
+            # if join is inner and it containes an empty df
+            # we return an empty df
+            return df.head(0)
+        if ignore_index:
+            df.index = o.index
+        else:
+            df.columns = result_columns.unique()
+        if not match_index:
             return df.sort_index()
         else:
-            return df
+            if sort:
+                return df.sort_index()
+            else:
+                return df
 
     typ = list(typs)[0]
 
@@ -259,24 +297,37 @@ def concat(objs, axis=0, ignore_index=False, sort=None):
             )
 
     if typ is cudf.DataFrame:
+        old_objs = objs
         objs = [obj for obj in objs if obj.shape != (0, 0)]
         if len(objs) == 0:
             # If objs is empty, that indicates all of
             # objs are empty dataframes.
             return cudf.DataFrame()
         elif len(objs) == 1:
-            if ignore_index:
-                result = cudf.DataFrame(
-                    data=objs[0]._data.copy(deep=True),
-                    index=cudf.RangeIndex(len(objs[0])),
-                )
+            if join == "inner":
+                data = None
             else:
-                result = objs[0].copy()
+                data = objs[0]._data.copy(deep=True)
+            result = cudf.DataFrame(
+                data=data,
+                index=cudf.RangeIndex(len(objs[0]))
+                if ignore_index
+                else objs[0].index.copy(deep=True),
+            )
             return result
         else:
-            return cudf.DataFrame._concat(
-                objs, axis=axis, ignore_index=ignore_index, sort=sort
+            if join == "inner" and len(old_objs) != len(objs):
+                # don't filter out empty df's
+                objs = old_objs
+            result = cudf.DataFrame._concat(
+                objs,
+                axis=axis,
+                join=join,
+                ignore_index=ignore_index,
+                sort=sort,
             )
+        return result
+
     elif typ is cudf.Series:
         objs = [obj for obj in objs if len(obj)]
         if len(objs) == 0: